Added knip to avoid unused imports and deps
