+++ package.json
@@ -56,7 +56,8 @@
     "test:pwa": "yarn workspace @conductor/pwa test",
     "test:scripts": "yarn workspace @conductor/scripts test",
     "test:extension": "yarn workspace @conductor/extension test",
-    "test:functions": "yarn workspace @conductor/functions test"
+    "test:functions": "yarn workspace @conductor/functions test",
+    "validate:knip": "knip"
   },
   "dependencies": {
     "firebase": "^11.2.0",
@@ -75,6 +76,7 @@
     "eslint-plugin-react-hooks": "^5.1.0",
     "globals": "^15.14.0",
     "jest": "^29.7.0",
+    "knip": "^5.44.4",
     "prettier": "^3.4.2",
     "ts-jest": "^29.2.5",
     "typescript": "^5.7.3",
